LayoutTests/crypto/subtle/aes-cfb-import-jwk-key-length-128-expected.txt

Test importing a JWK AES-CFB key with legnth 128

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CFB-8'
PASS key.algorithm.length is 128
PASS key.usages is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
